:wrench: move csv and add license
