+++ include/solvers/mpm_base.h
@@ -28,7 +28,7 @@
 #include "vector.h"
 
 // CSV-parser
-#include "csv.h"
+#include "csv/csv.h"
 
 namespace mpm {
